+++ src/app/about-us/about-us.component.css
@@ -1,7 +1,5 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
 
 h1 {
-  font-family: 'Roboto', sans-serif;
   font-weight: bolder;
 }
