+++ src/app/comments/comment-container.component.css
@@ -1,4 +1,4 @@
-h1 {
+h2 {
   border-bottom: 3px solid red;
   color: red;
 }
