+++ src/app/comments/comment-item/comment-item.component.html
@@ -1,10 +1,11 @@
 <mat-card>
   <mat-card-content>
-    {{commentMessage.body}}
+    {{comment.body}}
   </mat-card-content>
   <mat-card-subtitle>
-   {{commentMessage.name}} <br>
-   {{commentMessage.email}}
+    {{comment.name}}
   </mat-card-subtitle>
-
+  <mat-card-subtitle>
+    {{comment.email}}
+  </mat-card-subtitle>
 </mat-card>
