+++ src/app/header/header.component.css
@@ -1,17 +1,12 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
 
 .header{
- /* width: auto;*/
   background: white;
   padding: 20px;
   margin: 10px auto;
 }
 
 h1 {
-  font-family: 'Roboto', sans-serif;
   font-weight: bolder;
 }
 
-h2 {
-  font-family: 'Roboto', sans-serif;
-}
+
