+++ src/app/pipes/reverse.pipe.spec.ts
