+++ src/app/pipes/sum.pipe.ts
