+++ src/app/pipes/trucate.pipe.spec.ts
