+++ src/app/pipes/trucate.pipe.ts
