+++ src/app/post-details/post-details.component.css
@@ -1,13 +1,17 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
+.error {
+  color: red;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
 
 a {
   text-decoration: none;
   color: inherit;
-
 }
 
 h1 {
-  font-family: 'Roboto', sans-serif;
   font-weight: bolder;
 }
 
@@ -16,16 +20,13 @@ h2 {
   padding: 20px;
 }
 
-p {
-  font-family: 'Roboto', sans-serif;
-}
 
 img {
   max-width: 50%;
 }
 
 mat-card {
-  margin:  10px 0;
+  margin: 10px 0;
 }
 
 .header-mat-card {
@@ -36,5 +37,6 @@ mat-card {
 .body-mat-card {
   display: flex;
 }
+
