+++ src/app/post-details/post-details.component.ts
@@ -2,6 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute} from '@angular/router';
 import {Post} from '../posts/post.interface';
 import {PostsService} from 'src/app/posts/posts.service';
+
 import {switchMap, tap} from 'rxjs/operators';
 
 
@@ -29,21 +30,7 @@ export class PostDetailsComponent implements OnInit {
       .subscribe(response => {
         this.post = response;
       }, error => {
-        console.log('error', error);
         this.errorMessage = error;
       });
-
-
-    /* this.route.params.subscribe(params => {
-       this.id = params.id;
-       this.postsService.getPost(this.id)
-         .subscribe(post => {
-           this.post = post;
-         }, errorMsg => {
-           this.errorMessage = errorMsg;
-         });
-     });*/
-
   }
-
 }
