+++ src/app/posts/posts.module.ts
@@ -1,23 +1,24 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {PostsComponent} from './posts.component';
-import {HighlightPipe} from 'src/app/highlight.pipe';
-import {JoinPipe} from 'src/app/join.pipe';
-import {ReversePipe} from 'src/app/reverse.pipe';
-import {SumPipe} from 'src/app/sum.pipe';
-import {UcfirstPipe} from 'src/app/ucfirst.pipe';
+import {HighlightPipe} from 'src/app/pipes/highlight.pipe';
+import {JoinPipe} from 'src/app/pipes/join.pipe';
+import {ReversePipe} from 'src/app/pipes/reverse.pipe';
+import {SumPipe} from 'src/app/pipes/sum.pipe';
+import {UcfirstPipe} from 'src/app/pipes/ucfirst.pipe';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {MatCardModule} from '@angular/material/card';
 import {PostItemComponent} from 'src/app/posts/post-item/post-item.component';
-import {TrucatePipe} from 'src/app/trucate.pipe';
+import {TrucatePipe} from 'src/app/pipes/trucate.pipe';
 import {HttpClientModule} from '@angular/common/http';
-import {HTTP_INTERCEPTORS} from '@angular/common/http';
-import {HttpErrorInterceptor} from './http-error.interceptor';
 import {RouterModule} from '@angular/router';
 import {HeaderComponent} from 'src/app/header/header.component';
 import {PostDetailsComponent} from 'src/app/post-details/post-details.component';
 import { PostRoutingModule } from './post-routing.module';
-import { CommentModule } from '../post-details/comments/comment.module';
+import {CommentContainerComponent} from '../comments/comment-container.component';
+import {CommentModule} from '../comments/comment.module';
+
+
 
 @NgModule({
   declarations: [
@@ -31,6 +32,7 @@ import { CommentModule } from '../post-details/comments/comment.module';
     UcfirstPipe,
     PostItemComponent,
     TrucatePipe,
+    CommentContainerComponent
   ],
   exports: [
     PostsComponent
@@ -42,15 +44,9 @@ import { CommentModule } from '../post-details/comments/comment.module';
     HttpClientModule,
     RouterModule,
     PostRoutingModule,
-    CommentModule
-  ],
-  providers: [
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: HttpErrorInterceptor,
-      multi: true
-    }
+    CommentModule,
   ]
+
 })
 export class PostsModule {
 }
