+++ src/styles.css
@@ -1,12 +1,12 @@
-/* You can add global styles to this file, and also import other style files */
+@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
 
 html, body {
   height: 100vh;
 }
 
 body {
+  font-family: 'Roboto', sans-serif;
   max-width: 1100px;
-  font-family: Roboto, "Helvetica Neue", sans-serif;
   background: lightgray;
   margin: 0 auto;
 }
